[feat] 홈 카드 클릭 시 레이아웃 작업
홈 메인 페이지에서 카드 클릭 시 레이아웃 작업 완료

Related to : #25
