--- app/src/main/java/meaning/morning/HomeCardData.kt
@@ -1,2 +1,0 @@
-package meaning.morning
-
