--- app/src/main/java/meaning/morning/HomeFragment.kt
@@ -1,19 +1,0 @@
-package meaning.morning
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-
-class HomeFragment : Fragment() {
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? {
-
-        return inflater.inflate(R.layout.fragment_home, container, false)
-    }
-
-
-}
