+++ app/src/main/java/meaning/morning/presentation/CardTimeStampActivity.kt
@@ -1,0 +1,12 @@
+package meaning.morning.presentation
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import meaning.morning.R
+
+class CardTimeStampActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_card_time_stamp)
+    }
+}
