+++ app/src/main/java/meaning/morning/presentation/CardWriteDiaryActivity.kt
@@ -1,0 +1,12 @@
+package meaning.morning.presentation
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import meaning.morning.R
+
+class CardWriteDiaryActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_card_write_diary)
+    }
+}
