+++ app/src/main/java/meaning/morning/presentation/GroupFragment.kt
@@ -1,0 +1,20 @@
+package meaning.morning.presentation
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import meaning.morning.R
+
+
+class GroupFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_group, container, false)
+    }
+
+}
