+++ app/src/main/java/meaning/morning/presentation/HomeFragment.kt
@@ -1,0 +1,20 @@
+package meaning.morning.presentation
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import meaning.morning.R
+
+
+class HomeFragment : Fragment() {
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
+                              savedInstanceState: Bundle?): View? {
+
+        return inflater.inflate(R.layout.fragment_home, container, false)
+    }
+
+
+}
