+++ app/src/main/java/meaning/morning/presentation/data/HomeCardData.kt
@@ -1,0 +1,2 @@
+package meaning.morning
+
