+++ app/src/main/res/color/bottom_icon_color.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:state_checked="true" android:color="@color/meaning_navy"/>
-    <item android:color="@color/grey4"/>
+    <item android:state_checked="true" android:color="@color/meaning_navy_17"/>
+    <item android:color="@color/grey4_BD"/>
 </selector>
