+++ app/src/main/res/drawable/card_detail_button.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_focused="false" android:state_pressed="false">
+        <shape>
+            <solid android:color="@color/meaning_navy_17"/>
+            <corners android:radius="15dp" />
+        </shape>
+    </item>
+</selector>
