+++ app/src/main/res/drawable/card_promise_text.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_focused="false" android:state_pressed="false">
+        <shape>
+            <solid android:color="@color/meaning_white_FF"/>
+            <corners android:radius="10dp" />
+        </shape>
+    </item>
+</selector>
