+++ app/src/main/res/layout/activity_card_write_diary.xml
@@ -1,0 +1,126 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@color/meaning_lightblue_DE"
+    tools:context=".presentation.CardWriteDiaryActivity">
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/ct_topbar"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent">
+
+        <ImageView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:src="@drawable/back_btn"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            android:layout_marginStart="5dp"/>
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:includeFontPadding="false"
+            android:fontFamily="@font/notosanskr_medium"
+            android:text="하루 다짐"
+            android:layout_marginTop="4dp"
+            android:textColor="@color/grey1_33"
+            android:textSize="20sp"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+    <TextView
+        android:id="@+id/textView1"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="52dp"
+        android:fontFamily="@font/notosanskr_regular"
+        android:includeFontPadding="false"
+        android:text="오늘 아침을 글로 남겨봐요"
+        android:textColor="@color/grey2_4F"
+        android:textSize="18sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/ct_topbar" />
+
+    <EditText
+        android:id="@+id/et_diary"
+        android:layout_width="0dp"
+        android:layout_height="0dp"
+        android:layout_marginTop="84dp"
+        android:layout_marginBottom="109dp"
+        android:layout_marginStart="16dp"
+        android:layout_marginEnd="16dp"
+        android:paddingLeft="12dp"
+        android:paddingRight="12dp"
+        android:paddingTop="12dp"
+        android:gravity="top"
+        android:hint="나만 볼 수 있는 자기 회고 및 감사 일기를 써보세요!\n기분 좋은 아침을 시작하게 될 거에요"
+        android:textSize="17sp"
+        android:fontFamily="@font/spoqahansansneo_regular"
+        android:lineSpacingExtra="5dp"
+        android:maxLength="200"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/textView1"
+        app:layout_constraintBottom_toTopOf="@id/btn_register"
+        android:background="@drawable/card_promise_text"/>
+
+    <TextView
+        android:id="@+id/tv_inputNum"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintBottom_toTopOf="@id/btn_register"
+        app:layout_constraintTop_toBottomOf="@id/et_diary"
+        app:layout_constraintEnd_toEndOf="@id/tv_maxNum"
+        android:layout_marginTop="9dp"
+        android:layout_marginBottom="67dp"
+        android:layout_marginStart="309dp"/>
+
+    <TextView
+        android:id="@+id/tv_maxNum"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="/ 200"
+        app:layout_constraintStart_toEndOf="@id/tv_inputNum"
+        app:layout_constraintTop_toBottomOf="@id/et_diary"
+        app:layout_constraintBottom_toTopOf="@id/btn_register"
+        app:layout_constraintEnd_toEndOf="parent"
+        android:layout_marginTop="9dp"
+        android:layout_marginBottom="67dp"
+        android:layout_marginEnd="16dp"/>
+
+
+
+    <TextView
+        android:id="@+id/btn_register"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="16dp"
+        android:layout_marginEnd="16dp"
+        android:layout_marginBottom="99dp"
+        android:background="@drawable/card_detail_button"
+        android:fontFamily="@font/spoqahansansneo_regular"
+        android:paddingTop="16dp"
+        android:paddingBottom="16dp"
+        android:text="등록하기"
+        android:textAlignment="center"
+        android:textColor="@color/meaning_white_FF"
+        android:textSize="16sp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
+
+
+</androidx.constraintlayout.widget.ConstraintLayout>
