+++ app/src/main/res/layout/activity_main.xml
@@ -7,13 +7,13 @@
 
         <variable
             name="mainActivity"
-            type="meaning.morning.MainActivity" />
+            type="meaning.morning.presentation.MainActivity" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        tools:context=".presentation.MainActivity">
 
         <FrameLayout
             android:id="@+id/frame_layout"
