+++ app/src/main/res/layout/fragment_home.xml
@@ -3,9 +3,9 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/meaning_ivory"
+    android:background="@color/meaning_ivory_F6"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".HomeFragment">
+    tools:context=".presentation.HomeFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/ct_topbar"
@@ -71,7 +71,7 @@
             android:layout_marginTop="22dp"
             android:fontFamily="@font/spoqahansansneo_bold"
             android:text="좋은 아침입니다!"
-            android:textColor="@color/meaning_navy"
+            android:textColor="@color/meaning_navy_17"
             android:textSize="24sp"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
@@ -83,7 +83,7 @@
             android:layout_marginTop="10dp"
             android:fontFamily="@font/spoqahansansneo_thin"
             android:text="순서대로 기상미션을"
-            android:textColor="@color/meaning_navy"
+            android:textColor="@color/meaning_navy_17"
             android:textSize="24sp"
             app:layout_constraintStart_toStartOf="@+id/textView1"
             app:layout_constraintTop_toBottomOf="@+id/textView1" />
@@ -95,7 +95,7 @@
             android:layout_marginTop="10dp"
             android:fontFamily="@font/spoqahansansneo_thin"
             android:text="수행해볼까요?"
-            android:textColor="@color/meaning_navy"
+            android:textColor="@color/meaning_navy_17"
             android:textSize="24sp"
             app:layout_constraintStart_toStartOf="@+id/textView2"
             app:layout_constraintTop_toBottomOf="@+id/textView2" />
